+++ src/main/java/vboled/netcracker/musicstreamer/controllers/AdminController.java
@@ -6,11 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import vboled.netcracker.musicstreamer.model.user.User;
-import vboled.netcracker.musicstreamer.model.user.UserAdminView;
 import vboled.netcracker.musicstreamer.service.UserService;
-
-
-import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
